+++ src/pages/orders/order-page3.vue
@@ -1,13 +1,13 @@
 <template>
     <page-container>
-        <split-container mode="horizontal">
+        <!-- <split-container mode="column">
             <template v-slot:top>aaa</template>
             <template v-slot:bottom>bbb</template>
-        </split-container>
-        <!-- <split-container mode="vertical">
+        </split-container> -->
+        <split-container mode="row">
             <template v-slot:left>aaa</template>
             <template v-slot:right>bbb</template>
-        </split-container> -->
+        </split-container>
     </page-container>
 </template>